feat: 채팅 나가기 아이콘 수정
