+++ frontend/src/components/TopAppBar.tsx
@@ -6,6 +6,7 @@ import Typography from '../styles/Typography';
 import ArrowLeftIcon from '/src/assets/icons/arrow-left.svg?react';
 import LogoIcon from '/src/assets/icons/logo.svg?react';
 import LogoutIcon from '/src/assets/icons/logout.svg?react';
+import ExitIcon from '/src/assets/icons/exit.svg?react';
 import InfomationIcon from '/src/assets/icons/information-circle-contained.svg?react';
 import { getBorderStyle } from './border/getBorderType';
 
@@ -13,6 +14,7 @@ const IconComponents: { [key: string]: React.FC<React.SVGProps<SVGSVGElement>> }
   '/src/assets/icons/arrow-left.svg': ArrowLeftIcon,
   '/src/assets/icons/logo.svg': LogoIcon,
   '/src/assets/icons/logout.svg': LogoutIcon,
+  '/src/assets/icons/exit.svg': ExitIcon,
   '/src/assets/icons/information-circle-contained.svg': InfomationIcon,
 };
 
@@ -41,7 +43,6 @@ export const TopAppBar: React.FC<TopAppBarProps> = ({
 }) => {
   const LeftIconComponent = leftIconSrc ? IconComponents[leftIconSrc] : null;
   const RightIconComponent = rightIconSrc ? IconComponents[rightIconSrc] : null;
-  
   return (
     <TopAppBarContainer backgroundColor={backgroundColor} border={border}>
       {LeftIconComponent && (
